Сделал рабочей кнопку показать больше
