+++ src/main.js
@@ -1,13 +1,3 @@
-import SiteMenuView from "./view/site-menu/site-menu";
-import ProfileView from "./view/profile/profile";
-import SortView from "./view/sort/sort";
-import ShowMoreButtonView from "./view/show-more-button/show-more-button";
-import FilmsCountView from "./view/films-count/films-count";
-import FilmsWrapperView from "./view/films/films";
-import FilmsListView from "./view/films-list/films-list";
-import EmptyFilmsListView from "./view/films-list-empty/films-list-empty";
-import FilmListContainer from "./view/films-list-container/films-list-container";
-import FilmCardView from "./view/film-card/film-card";
 import FilmPopupView from "./view/film-popup/film-popup";
 import ClosePopupButtonWrapperView from "./view/close-popup-button-wrapper/close-popup-button-wrapper";
 import ClosePopupButtonView from "./view/close-popup-button/close-popup-button";
@@ -21,21 +11,11 @@ import FilmPopupControlsView from "./view/film-popup-controls/film-popup-control
 import FilmPopupCommentsWrapView from "./view/film-popup-comments-wrap/film-popup-comments-wrap";
 import FilmPopupCommentsListView from "./view/film-popup-comments-list/film-popup-comments-list";
 import FilmPopupNewCommentView from "./view/film-popup-new-comment/film-popup-new-comment";
-import {render, remove, RenderPosition} from "./utils/render";
-import {isEscEvent, getRandomIntInRange} from "./utils/common";
+import {render, remove} from "./utils/render";
+import {isEscEvent} from "./utils/common";
 import {generateFilmCards} from "./mock/film-card";
 import {generateComments} from "./mock/comment";
-
-const CARDS_TO_SHOW_COUNT = 5;
-
-const MIN_PROFILE_RANK = 0;
-const MAX_PROFILE_RANK = 40;
-
-const ELEMENTS_TO_SHOW_POPUP = [
-  `film-card__poster`,
-  `film-card__title`,
-  `film-card__comments`
-];
+import FilmListPresenter from "./presenter/film-list";
 
 const allFilmcards = generateFilmCards();
 
@@ -52,69 +32,6 @@ const siteBodyElement = document.querySelector(`body`);
 const siteHeaderElement = siteBodyElement.querySelector(`.header`);
 const siteMainElement = siteBodyElement.querySelector(`.main`);
 const siteFooterElement = siteBodyElement.querySelector(`.footer`);
-
-const sortComponent = new SortView();
-const profileComponent = new ProfileView(getRandomIntInRange(MAX_PROFILE_RANK, MIN_PROFILE_RANK));
-const siteMenuComponent = new SiteMenuView(allFilmcards);
-
-render(siteHeaderElement, profileComponent);
-render(siteMainElement, siteMenuComponent, RenderPosition.AFTERBEGIN);
-render(siteMainElement, sortComponent);
-
-const filmsWrapperComponent = new FilmsWrapperView();
-const filmsListComponent = new FilmsListView();
-const emptyFilmsList = new EmptyFilmsListView();
-const filmsListContainerComponent = new FilmListContainer();
-
-render(siteMainElement, filmsWrapperComponent);
-
-if (allFilmcards.length > 0) {
-  render(filmsWrapperComponent, filmsListComponent);
-} else {
-  remove(sortComponent);
-  render(filmsWrapperComponent, emptyFilmsList);
-}
-
-render(filmsListComponent, filmsListContainerComponent);
-
-const showMoreButtonComponent = new ShowMoreButtonView();
-render(filmsListComponent, showMoreButtonComponent);
-
-const renderFilmCards = (cardsToShow) => {
-  cardsToShow.forEach((cardToShow) => {
-    const commentsCount = getFilmCardComments(cardToShow).length;
-    const filmCardComponent = new FilmCardView(cardToShow, commentsCount);
-    render(filmsListContainerComponent, filmCardComponent);
-  });
-};
-
-const appendFilmCards = (cardsToShow) => {
-  filmsListContainerComponent.element.innerHTML = ``;
-
-  renderFilmCards(cardsToShow);
-
-  if (allFilmcards.length === cardsToShow.length) {
-    remove(showMoreButtonComponent);
-  }
-};
-
-const initialCardsToShow = allFilmcards.slice(0, CARDS_TO_SHOW_COUNT);
-appendFilmCards(initialCardsToShow);
-
-let showMoreButtonClickCounter = 1;
-
-const onShowMoreButtonClick = () => {
-  showMoreButtonClickCounter++;
-
-  const cardsToShow = allFilmcards.slice(0, CARDS_TO_SHOW_COUNT * showMoreButtonClickCounter);
-
-  appendFilmCards(cardsToShow);
-};
-
-showMoreButtonComponent.setClickHandler(onShowMoreButtonClick);
-
-const filmsCountComponent = new FilmsCountView(allFilmcards.length);
-render(siteFooterElement, filmsCountComponent);
 
 const filmPopupComponent = new FilmPopupView();
 const closePopupButtonWrapperComponent = new ClosePopupButtonWrapperView();
@@ -188,25 +105,5 @@ const renderPopup = (card) => {
   siteBodyElement.classList.toggle(`hide-overflow`);
 };
 
-const isPopupElementClicked = (className) => ELEMENTS_TO_SHOW_POPUP.some((val) => val === className);
-
-const getCard = (id) => allFilmcards.find((el) => el.id === id);
-
-const onFilmCardClick = (evt) => {
-  const showPopup = isPopupElementClicked(evt.target.className);
-
-  if (showPopup) {
-    evt.preventDefault();
-
-    const cardId = evt.target.parentNode.dataset.id;
-    const clickedCard = getCard(cardId);
-
-    if (clickedCard) {
-      renderPopup(clickedCard);
-      document.addEventListener(`keydown`, onPopupEscPress);
-    }
-  }
-};
-
-filmsListContainerComponent.setClickHandler(onFilmCardClick);
-
+const filmListPresenter = new FilmListPresenter(siteHeaderElement, siteMainElement, siteFooterElement);
+filmListPresenter.init(allFilmcards, allComments);
